requirements.txt
certifi==2022.9.24
charset-normalizer==2.1.1
greenlet==2.0.1
hvac==1.0.2
idna==3.4
psycopg2==2.9.5
pyhcl==0.4.4
python-dotenv==0.21.0
requests==2.28.1
SQLAlchemy==1.4.44
urllib3==1.26.12
